added locale switcher & localized more texts.
