+++ src/components/AppMethodPrimary.vue
@@ -1,32 +1,32 @@
 <template>
   <div>
-    <label for="firstmethod">I have an array, I would like to</label>
+    <label for="firstmethod">{{$t('firstMethod', 'I have an array, I would like to')}}</label>
     <select id="firstmethod" name="select" v-model="selectedFilter">
       <option value="" disabled selected>...</option>
       <option v-for="option in options" :key="option.name" :value="option">
         {{ option }}
       </option>
     </select>
-    <div v-if="selectedFilter === 'add items or other arrays'">
+    <div v-if="selectedFilter === options[0]">
       <method-choice :options="adding" methodType="add"/>
     </div>
-    <div v-else-if="selectedFilter === 'remove items'">
+    <div v-else-if="selectedFilter === options[1]">
       <method-choice :options="removing" methodType="remove"/>
     </div>
-    <div v-else-if="selectedFilter === 'walk over items'">
+    <div v-else-if="selectedFilter === options[3]">
       <method-choice :options="iterate" methodType="iterate by"/>
     </div>
-    <div v-else-if="selectedFilter === 'find items'">
+    <div v-else-if="selectedFilter === options[2]">
 
       <p>
-        <label for="findmethod">I'm trying to find</label>
+        <label for="findmethod">{{$t('findMethod', "I'm trying to find")}}</label>
         <select id="findMethod" name="select" v-model="selectedFind">
           <option value="" disabled selected>...</option>
-          <option value="single">one item</option>
-          <option value="many">one or many items</option>
+          <option value="single">{{$t('singleItem', 'one item')}}</option>
+          <option value="many">{{$t('manyItems', 'one or many items')}}</option>
         </select>
       </p>
-      
+
       <div v-if="selectedFind === 'single'">
         <method-choice :options="find.single" methodType="find"/>
       </div>
@@ -36,21 +36,47 @@
       </div>
 
     </div>
-    <div v-else-if="selectedFilter === 'order an array'">
+    <div v-else-if="selectedFilter === options[5]">
       <method-choice :options="ordering"/>
     </div>
-    <div v-else-if="selectedFilter === 'return a string'">
+    <div v-else-if="selectedFilter === options[4]">
       <method-choice :options="string"/>
     </div>
-    <div v-else-if="selectedFilter === 'something else'">
+    <div v-else-if="selectedFilter === options[6]">
       <method-choice :options="other"/>
     </div>
   </div>
 </template>
 
 <script>
+import {store} from '../../store/index'
 import { mapState } from 'vuex'
 import MethodChoice from './methods/MethodChoice.vue'
+
+/*function localizedState({$store}, prop) {
+  const curLang = $store.state.curLanguage
+  console.log('state', curLang);
+  return $store.state[curLang][prop];
+}*/
+
+function mapLocalizedState(props) {
+  let res = {};
+  props.forEach(prop => res[prop] = (state) => {
+    const {curLanguage} = state.$store.state;
+    return state.$store.state[curLanguage][prop];
+  });
+  return res;
+}
+
+// const primaryOptionsDefault = [ // default fallback
+//   'add items or other arrays',
+//   'remove items',
+//   'find items',
+//   'walk over items',
+//   'return a string',
+//   'order an array',
+//   'something else'
+// ];
 
 export default {
   components: {
@@ -60,19 +86,23 @@ export default {
     return {
       selectedFilter: '',
       selectedFind: '',
-      options: [
-        'add items or other arrays',
-        'remove items',
-        'find items',
-        'walk over items',
-        'return a string',
-        'order an array',
-        'something else'
-      ]
     }
   },
   computed: {
-    ...mapState([
+    ...mapLocalizedState([
+      'adding', //: state => localizedState(state, 'adding'),
+      'removing',
+      'iterate',
+      'string',
+      'ordering',
+      'other',
+      'find'
+    ]),
+    options: function() {
+      return this.$t('primaryOptions');
+    }
+  },
+    /*...mapState([
       'adding',
       'removing',
       'iterate',
@@ -80,8 +110,8 @@ export default {
       'ordering',
       'other',
       'find'
-    ])
-  },
+    ].map(prop => `${store.getters.curLanguage}.${prop}`))*/
+  // },
   watch: {
     selectedFilter() {
       this.$store.commit('resetSelection')
