+++ src/locale/ar/index.js
@@ -1,0 +1,24 @@
+export default {
+  firstMethod: 'I have an array, I would like to',
+  findMethod: "I'm trying to find",
+  methodOptions: 'I need to',
+  methodTypes: {
+    add: 'add',
+    remove: 'remove',
+    find: 'find',
+    'iterate by': 'iterate by',
+    'string': 'string'
+  },
+  singleItem: 'one item',
+  manyItems: 'one or many items',
+  primaryOptions: [
+    'add items or other arrays',
+    'remove items',
+    'find items',
+    'walk over items',
+    'return a string',
+    'order an array',
+    'something else'
+  ]
+  // other text can be added here
+}
