+++ src/locale/de/index.js
@@ -1,0 +1,24 @@
+export default {
+  firstMethod: 'Ich habe ein Array und würde gerne', //'I have an array, I would like to',
+  findMethod: 'Ich möchte folgendes finden', // "I'm trying to find"
+  methodOptions: 'Ich muss',
+  methodTypes: {
+    add: 'hinzufügen',
+    remove: 'entfernen',
+    find: 'finden',
+    'iterate by': 'durchlaufen',
+    'string': 'String zurückgeben'
+  },
+  singleItem: 'ein Element', // 'one item',
+  manyItems: 'ein oder mehrere Elemente', // 'one or many items',
+  primaryOptions: [
+    'Elemente oder Arrays hinzufügen', // ich würde gerne i would like to ... add items or other arrays
+    'Elemente entfernen',
+    'Elemente finden',
+    'Elemente durchlaufen',
+    'einen String zurückgeben',
+    'ein Array sortieren',
+    'etwas anders machen'
+  ]
+  // other text can be added here
+}
