+++ store/ar/meta.json
@@ -1,0 +1,5 @@
+{
+  "short": "ar",
+  "long": "Arabic",
+  "createdBy": ""
+}
