+++ store/de/meta.json
@@ -1,0 +1,5 @@
+{
+  "short": "de",
+  "long": "German",
+  "createdBy": ""
+}
