+++ store/en/meta.json
@@ -1,0 +1,5 @@
+{
+  "short": "en",
+  "long": "English",
+  "createdBy": "sdras"
+}
