+++ store/es/meta.json
@@ -1,0 +1,5 @@
+{
+  "short": "es",
+  "long": "Spanish",
+  "createdBy": ""
+}
