+++ store/fr/index.js
@@ -1,9 +1,4 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
-
-Vue.use(Vuex)
-
-export const store = new Vuex.Store({
+export default {
   state: {
     selectedMethod: '',
     adding: [
@@ -261,13 +256,5 @@ export const store = new Vuex.Store({
         }
       ]
     }
-  },
-  mutations: {
-    selectionMethod(state, selected) {
-      state.selectedMethod = selected
-    },
-    resetSelection(state) {
-      state.selectedMethod = ''
-    }
   }
-})
+}
