+++ store/fr/meta.json
@@ -1,0 +1,5 @@
+{
+  "short": "fr",
+  "long": "French",
+  "createdBy": ""
+}
